reorg this a bit and some cleanup
I experimented too with an alternative SSE2 implementation to see if it was any better than .NET's. Seemed to be barely 10%ish faster, which isn't really much in the grand scheme of things.
