+++ ExternalProjects/LibBizHash/Makefile
@@ -1,9 +1,9 @@
 CC = gcc
 
-CFLAGS = -Wall -Wextra -O3 -flto -s -fvisibility=internal -fPIC
-LFLAGS = -shared
+CFLAGS = -Wall -Wextra -O3 -flto -fvisibility=internal -fPIC -Icommon
+LFLAGS = -s -shared
 
-SRCS = $(wildcard ./*.c) $(wildcard x86/*.c)
+SRCS = $(wildcard common/*.c) $(wildcard crc32/*.c) $(wildcard sha1/*.c) bizinterface.c
 
 ifeq ($(OS),Windows_NT)
 EXT = dll
