+++ ExternalProjects/LibBizHash/bizinterface.c
@@ -1,25 +1,26 @@
 #include <stdbool.h>
-#include <string.h>
-#include "cpu_features.h"
+#include "common.h"
 
 __attribute__((visibility("default")))
 crc32_func BizCalcCrcFunc(void) {
-	cpu_check_features();
+	x86_check_features();
 	return (x86_cpu_has_pclmulqdq && x86_cpu_has_sse41) ? &crc32_pclmulqdq : &crc32_braid;
 }
 
 __attribute__((visibility("default")))
 bool BizSupportsShaInstructions(void) {
-	cpu_check_features();
+	x86_check_features();
 	return x86_cpu_has_sha && x86_cpu_has_sse41;
 }
 
 __attribute__((visibility("default")))
-void BizFastCalcSha1(uint32_t state[5], const uint8_t data[], uint32_t length) {
-	uint64_t bit_length = length * 8;
+void BizCalcSha1(uint32_t state[5], const uint8_t data[], uint32_t length) {
+	x86_check_features();
+
+	uint64_t bit_length = length * 8ULL;
 
 	// hash most of the data, leaving at most 63 bytes left
-	BizFastCalcSha1Internal(state, data, length);
+	sha1_sha(state, data, length);
 	data += length & ~0x3F;
 	length &= 0x3F;
 
@@ -33,7 +34,7 @@ void BizFastCalcSha1(uint32_t state[5], const uint8_t data[], uint32_t length) {
 	// the last 8 bytes in the last block contain the data length;
 	// if the current block is too full hash it and start a new one (here the old one is cleared and re-used)
 	if (__builtin_expect(length > 56, false)) {
-		BizFastCalcSha1Internal(state, block, 64);
+		sha1_sha(state, block, 64);
 		memset(block, 0, 56);
 	}
 
@@ -42,5 +43,12 @@ void BizFastCalcSha1(uint32_t state[5], const uint8_t data[], uint32_t length) {
 		block[63 - i] = bit_length >> i * 8;
 	}
 	// hash the last block
-	BizFastCalcSha1Internal(state, block, 64);
+	sha1_sha(state, block, 64);
+
+	// byteswap state (to big endian format)
+	state[0] = __builtin_bswap32(state[0]);
+	state[1] = __builtin_bswap32(state[1]);
+	state[2] = __builtin_bswap32(state[2]);
+	state[3] = __builtin_bswap32(state[3]);
+	state[4] = __builtin_bswap32(state[4]);
 }
