+++ src/BizHawk.Common/checksums/LibBizHash.cs
@@ -17,6 +17,6 @@ namespace BizHawk.Common
 		public static extern bool BizSupportsShaInstructions();
 
 		[DllImport("libbizhash", CallingConvention = cc)]
-		public static extern void BizFastCalcSha1(IntPtr state, byte[] data, int len);
+		public static extern void BizCalcSha1(IntPtr state, byte[] data, int len);
 	}
 }
